IT WORKS!
Match3.java is the Driver- it sets up the main class and creates the
first state.
The First state creates a CardHandler with update(), render(), and
recursive logic.  Cards now contain images.

The images shouldn't be saved directly in the Card class, instead the
Card class should invoke an animation Object.
I have uploaded this as a java project, simply add it to eclipse, add
the jars (see the wiki on how to add necessary jars) and run the game.